samples/python-validator/requirements.txt
openai>=1.54.0
azure-identity>=1.19.0
